Move AppData into kas_core::app
