+++ crates/kas-core/src/app/app.rs
@@ -5,13 +5,13 @@
 
 //! [`Application`] and supporting elements
 
-use super::{AppGraphicsBuilder, AppState, Platform, ProxyAction, Result};
+use super::{AppData, AppGraphicsBuilder, AppState, Platform, ProxyAction, Result};
 use crate::config::Options;
 use crate::draw::{DrawShared, DrawSharedImpl};
 use crate::event;
 use crate::theme::{self, Theme, ThemeConfig};
 use crate::util::warn_about_error;
-use crate::{impl_scope, AppData, Window, WindowId};
+use crate::{impl_scope, Window, WindowId};
 use std::cell::RefCell;
 use std::rc::Rc;
 use winit::event_loop::{EventLoop, EventLoopBuilder, EventLoopProxy};
