+++ crates/kas-core/src/lib.rs
@@ -31,7 +31,7 @@ mod root;
 pub use crate::core::*;
 pub use action::Action;
 pub use decorations::Decorations;
-pub use erased::{AppData, Erased, ErasedStack};
+pub use erased::{Erased, ErasedStack};
 pub use kas_macros::*;
 #[doc(inline)] pub use popup::Popup;
 #[doc(inline)] pub(crate) use popup::PopupDescriptor;
